pkg/cvo/sync_worker: Make expected/actual version mismatch fatal
If, for example, the configured ClusterVersion spec.upstream
advertised a given image as 4.0.1, but the version baked into the
release's own metadata was 1.0.0-abc, report
VerifyPayloadVersionFailed and continue to apply the previous release
image, instead of optimistically moving on to apply the new release
image.  This protects users from compromised or man-in-the-middled
upstreams who attempt downgrade and similar attacks by misrepresenting
a recommended update.

Addressing a FIXME from 58356deee6 (*: Port from Update to Release for
ClusterVersion status, 2020-07-24, #419).