pkg/cvo/sync_worker: TaskNodeComplete events during updates
Only during updates, because:

* Install-time is a free-for-all, where the CVO doesn't block on
  anything.  This would be a lot of "node complete" noise about nodes
  where we had only attempted to push manifests, and that's unlikely
  to be what event-readers expect TaskNodeComplete to imply.

* Reconcile-time hopefully has very few instances where the CVO needs
  to stomp on changes, block on a recently Available=False operator,
  etc.  Eventing on each completed TaskNode would be lots of noise
  without much interesting signal.

During updates, we have the structured graph and blocking TaskNodes
described in docs/user/reconciliation.md, and the flow through that
graph is what the events from this commit will help shed light on.
You could also achieve this by preserving logs from the CVO pods as
they are repositioned throughout an update, but we don't have tooling
in CI to do that conveniently today.

The hardcoded name and namespace for cvoObjectRef isn't great (for
example, it won't work in pkg/start/start_integration_test.go , where
the ClusterVersion's name and namespace are random).  But it's the
pattern we've used since we started eventing in 475e71f334 (emit
events for each new payload, 2020-07-21, #411), so I'm recylcing it
for now.

Also log ApplyFailed events when we fail in apply, to remove some of
the guesswork in determining what manifest(s) had trouble.