pkg/cvo: Separate ConfigMap informer for openshift-config-managed
With c9fab435c1 (pkg/cvo: Fetch proxy CA certs from
openshift-config-managed/trusted-ca-bundle, 2020-01-31, #311), I
attempted to pivot to loading the TLS configration from the
openshift-config-managed namespace.  But I hadn't realized that
cmConfigInformer was scoped to the openshift-config namespace, so
attempts to retrieve trusted-ca-bundle failed via the "not found"
no-op path regardless of whether the ConfigMap actually existed.  With
this commit, I create a new informer for the openshift-config-managed
namespace, so we can successfully retrieve the ConfigMap when it
exists.

An alternative approach would be to drop the informers.WithNamespace
filter.  My impression is that having two namespace-filtered informers
will be more efficient than a single unfiltered informer, but I'm not
really sure.