pkg/cvo/metrics: Abandon child goroutines after shutdownContext expires
Apparently there's something in the HTTPS server goroutine that can
hang up even if we've called Shutdown() on the server [1].  Defend
against that with a safety valve to abandon stuck goroutines if
shutdownContext expires.

Also pivot to resultChannel and asyncResult, so we can get names for
the collected channels (and more easily identify the stuck channels by
elimination), following the pattern set by cc1921d186 (pkg/start:
Release leader lease on graceful shutdown, 2020-08-03, #424).

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=1891143#c1