docs/dev/clusteroperator: Fix "he" -> "the" typo
This snuck in with be09d1e4d3 (Doc fix for cluster operator creation
step, 2020-10-30, #478).

Re: "or recreate if deleted", the CVO (if it is running) will
eventually get around to recreating the ClusterOperator, as described
in the text added in be09d1e4d3.  But it's still the operator's job to
recreate it if the ClusterOperator is missing, as part of the
"properly update" responsibility.