pkg/cvo/internal: Do not block on Degraded=True ClusterOperator
We have blocking on this condition since 545c3424aa (api: make status
substruct on operatorstatus, 2018-10-15, #31) when it was Failing.
We'd softened our install-time handling to act this way back in
b0b4902fce (clusteroperator: Don't block on failing during
initialization, 2019-03-11, #136), motivated by install speed [1].
And a degraded operator may slow dependent components in their own
transitions.  But as long as the operator/operand are available at
all, it should not block depndent components from transitioning, so
this commit removes the Degraded=True block from the remaining modes.

We still have the warning ClusterOperatorDegraded alerting admins
when an operator goes Degraded=True for a while, we will just no
longer block updates at that point.  We won't block ReconcilingMode
manifest application either, but since that's already flattened and
permuted, and ClusterOperator tend to be towards the end of their
TaskNode, the impact on ReconcilingMode is minimal (except that we
will no longer go Failing=True in ClusterVersion when the only issue
is some Degraded=True ClusterOperator).

[1]: https://github.com/openshift/cluster-version-operator/pull/136#issue-260160975