pkg/cvo: Compare Cincinnati data by digest when merging metadata
Fixing [1], where clusters installed with pullspecs like:

  registry.svc.ci.openshift.org/ocp/release@sha256:c7e8f18e8116356701bd23ae3a23fb9892dd5ea66c8300662ef30563d7104f39

would not pick up channel metadata that Cincinnati was associating
with canonical pullspecs like:

  quay.io/openshift-release-dev/ocp-release@sha256:c7e8f18e8116356701bd23ae3a23fb9892dd5ea66c8300662ef30563d7104f39

I really wish source-repo "hints" were not part of pullspecs, and
these images were purely content-addressable.  But wishing doesn't
make it true ;).

The "Identical by tag" case is because even though we have no
guarantee that a particular tag is stable, in any given moment there
should only be one manifest backing a given pullspec, regardless of
whether that pullspec is by-tag or by-digest.  This is unlikely to
matter in the wild, where Cincinnati will serve by-digest pullspecs.
But it currently matters for CVO unit tests, where we rely on
image/image:v1.0.0 matching for metadata population.  And it matches
the:

  merged.Image == availableUpdates.Current.Image

and similar conditions we're replacing with equalDigest (so
equalDigest will match what we used to match, and also the
same-digest-different-repo cases we didn't match before).

I'm not using the new equalDigest for some of our "is this change an
update?" logic, because we want to go through "updates" on repository
changes [2] to ensure that the release content is still available when
accessed via the new pullspec (although the presence of
ImageContentSourcePolicies may mean that even with a change to the
repository portion of the pullspec, we still end up actually pulling
the release from the same location we were initially using).

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=1879976
[2]: https://bugzilla.redhat.com/show_bug.cgi?id=1879963#c4