pkg/cvo/cvo: Guard Operator.Run goroutine handling from early cancels
In [1], we had a CVO fail to gracefully release the leader lock, with:

  $ grep start.go: namespaces/openshift-cluster-version/pods/cluster-version-operator-968d9fd48-25zkz/cluster-version-operator/cluster-version-operator/logs/current.log
  2021-01-26T19:02:48.678276895Z I0126 19:02:48.678120       1 start.go:21] ClusterVersionOperator 4.6.0-202012050130.p0-39a4256
  2021-01-26T19:02:49.273117163Z I0126 19:02:49.273074       1 start.go:260] Waiting on 2 outstanding goroutines.
  2021-01-26T19:02:49.273512006Z E0126 19:02:49.273474       1 start.go:271] Collected metrics server goroutine: listen tcp 0.0.0.0:9099: bind: address already in use
  2021-01-26T19:02:49.273587622Z I0126 19:02:49.273551       1 start.go:260] Waiting on 1 outstanding goroutines.
  2021-01-26T19:02:49.273587622Z I0126 19:02:49.273569       1 start.go:264] Run context completed; beginning two-minute graceful shutdown period.
  2021-01-26T19:02:49.273587622Z I0126 19:02:49.273577       1 start.go:260] Waiting on 1 outstanding goroutines.
  2021-01-26T19:04:49.273938652Z I0126 19:04:49.273877       1 start.go:260] Waiting on 1 outstanding goroutines.
  2021-01-26T19:05:39.264746935Z I0126 19:05:39.264690       1 start.go:190] Shutting down due to terminated

So the bind failed, which should be rare, but can happen if the new
CVO is scheduled before the kernel gets around to reaping the port
from a previous CVO pod.  The start.go handling notices the metrics
server dying, and immediately cancels runContext, and we go into the
two-minute shutdown grace period.  That happened so early that the
context was already done by the time we got into Operator.Run, with
its:

  go wait.UntilWithContext(runContext, func(runContext context.Context) {
    defer close(workerStopCh)

From the godocs for UntilWithContext:

  $ git --no-pager grep -n '// UntilWithContext is syntactic sugar\|f may not be invoked if context\|func JitterUntilWithContext' vendor/k8s.io/apimachinery/pkg/util/wait
  vendor/k8s.io/apimachinery/pkg/util/wait/wait.go:95:// UntilWithContext is syntactic sugar on top of JitterUntilWithContext
  vendor/k8s.io/apimachinery/pkg/util/wait/wait.go:183:// Cancel context to stop. f may not be invoked if context is already expired.
  vendor/k8s.io/apimachinery/pkg/util/wait/wait.go:184:func JitterUntilWithContext(ctx context.Context, f func(context.Context), period time.Duration, jitterFactor float64, sliding bool) {

So when runContext is already done, we never close workerStopCh, and
we hang forever when we block on it before returning from
Operator.Run.

This commit addresses that issue by bringing in a resultChannel flow
like we already use in pkg/start/start.go and pkg/cvo/metrics.go.  It
will give us logging for the number of outstanding goroutines, and
tell us as each of them are reaped, which will make it easier to
diagnose any other issues with hanging children, while also using
shutdownContext to bail out to avoid a stuck child causing major
problems.

This commit also adds ShutDown calls for availableUpdatesQueue and
upgradeableQueue, so we don't hang those two goroutines in
processNextWorkItem's queue.Get() call.  The shutdowns should have
been added when the queues landed in 90e9881bd5 (cvo: Change the core
CVO loops to report status to ClusterVersion, 2018-11-02, #45) and
04528144fe (cvo: long running upgradeable sync worker, 2019-09-05,
until this commit tightened down on the goroutine collection.

There is a risk with these cluster-touching children, that if we hit
the shutdownContext-timeout safety valve, we might exit Operator.Run
and release the leader lock, and then have our abandoned child
fighting over in-cluster state with some subsequent CVO which just
acquired the leader lock.  We might want to write an event or
something to make it more obvious when that happens, but for now, I'm
more interested in closing the context-already-canceled bug.

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=1921413