pkg/cvo/internal/operatorstatus: Drop deprecated failing/progressing handling
Also simplify the 'undone' type from a map to a slice of strings.  The
only consumers are length checks and nestedMessage formatting, so we
don't need to preserve structured access to the current and expected
values.

I've also dropped the old, early "Cluster operator %s is still
updating" guard.  If the ClusterOperator is degraded or unavailable,
those are much more serious than being behind in version, so we want
those guards coming first.  The undone guard is now only the final
fallback for available, non-degraded operators that aren't completely
satisfactory (and I've pivoted to use ClusterOperatorUpdating for that
reason).

I'm also now keeping any available condition data, and including that
status, reason, and message in the nested message, because it seemed
strange to do that for degraded, but not for the more serious
unavailable.

And finally, I've dropped the old mode switch, and instead distributed
the mode check to each relevant guard.  The "we're completely
satisfied" case is now where you end up if none of the "we aren't
happy" guards trip.  This avoids the need to have an "if we got this
far, it must have been because we have undone versions" assumption for
the ClusterOperatorUpdating block.