pkg/cvo/metrics: Exclude unrecognized statuses from cluster_operator_conditions
We have been excluding Unknown since the metric landed in 8b9118992d
(cvo: Add prometheus metrics to the CVO for current update state,
2018-11-01, #45).  But operators could do crazy things like setting
'status: Muahaha', and we don't want to collapse them all into the
False/0 case.  We could enumerate the known statuses (e.g. using 2 for
Unknown), but that would probably not play well with existing queries,
and Unknown is hopefully not too popular.