pkg/cvo/sync_worker: Skip precreation of baremetal ClusterOperator
This is a hack fix for [1], where we have a delay on 4.6->4.7 updates,
and on some 4.7 installs, between the very early ClusterOperator
precreation and the operator eventually coming up to set its status
conditions.  In the interim, there are no conditions, which causes
cluster_operator_up to be 0, which causes the critical
ClusterOperatorDown to fire.  We'll want a more general fix going
forward, this commit is a temporary hack to avoid firing the critical
ClusterOperatorDown while we build consensus around the general fix.

The downside to dropping precreates for this operator is that we lose
the must-gather references when the operator fails to come up.  That
was what precreation was designed to address in 2a469e37c1 (cvo: When
installing or upgrading, fast-fill cluster-operators, 2020-02-07, #318).
If we actually get a must-gather without the bare-metal bits and we
miss them, we can revisit the approach this hack is taking.

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=1929917