docs/user/status: Mention channel-clearing for VersionNotFound
I would prefer to configure "don't poll the upstream" by clearing the
'upstream' setting.  But sadly, it seems that the CVO has been falling
back to a default URI when the ClusterVersion upstream is empty since
way back [1,2], and that this behavior is enshrined in the API [3].
Although the channel docs also talk about defaults [4], the only
channel defaulting in the CVO is when creating a ClusterVersion object
after the in-cluster copy was (accidentally?)  deleted [5].  So maybe
we could talk folks into adjusting the CVO logic to return NoUpstream
in the empty-upstream case, but at the moment, clearing the channel is
the best approach for "the CVO keeps complaining that it can't hit the
upstream, and I want ot quiet it down [6]".

[1]: https://github.com/openshift/cluster-version-operator/blame/2c4931dc283c551938be1a00fee290de0b79d99c/pkg/cvo/availableupdates.go#L27-L31
[2]: openshift/cluster-version-operator@ab4d84a#diff-78f2af341fa49292dd6930f378018867R24
[3]: https://github.com/openshift/api/blame/0422dc17083e9e8df18d029f3f34322e96e9c326/config/v1/types_cluster_version.go#L56-L57
[4]: https://github.com/openshift/api/blame/0422dc17083e9e8df18d029f3f34322e96e9c326/config/v1/types_cluster_version.go#L62-L63
[5]: https://github.com/openshift/cluster-version-operator/blob/2c4931dc283c551938be1a00fee290de0b79d99c/pkg/cvo/cvo.go#L602
[6]: https://bugzilla.redhat.com/show_bug.cgi?id=1827378