pkg/cvo/sync_worker: Shift ClusterOperator pre-creation into the manifest-task node
ClusterOperator pre-creation landed in 2a469e37c1 (cvo: When
installing or upgrading, fast-fill cluster-operators, 2020-02-07, #318)
to move us from:

1. CVO creates a namespace for an operator.
2. CVO creates ... for the operator.
3. CVO creates the operator Deployment.
4. Operator deployment never comes up, for whatever reason.
5. Admin must-gathers.
6. Must gather uses ClusterOperators for discovering important stuff,
   and because the ClusterOperator doesn't exist yet, we get no data
   about why the deployment didn't come up.

to:

1. CVO pre-creates ClusterOperator for an operator.
2. CVO creates the namespace for an operator.
3. CVO creates ... for the operator.
4. CVO creates the operator Deployment.
5. Operator deployment never comes up, for whatever reason.
6. Admin must-gathers.
7. Must gather uses ClusterOperators for discovering important stuff,
   and finds the one the CVO had pre-created with hard-coded
   relatedObjects, gathers stuff from the referenced operator
   namespace, and allows us to trouble-shoot the issue.

But when ClusterOperator pre-creation happens at the beginning of an
update sync cycle, it can take a while before the CVO gets from the
ClusterOperator creation in (1) to the operator managing that
ClusterOperator in (4), which can lead to ClusterOperatorDown alerts
[1,2].

fdef37dc0e (pkg/cvo/sync_worker: Skip precreation of baremetal
ClusterOperator, 2021-03-16, #531) landed a narrow hack to avoid
issues on 4.6 -> 4.7 updates, which added the baremetal operator [1].
But we're adding a cloud-controller-manager operator in 4.7 -> 4.8,
and breaking the same way [2].  This commit pivots to a more generic
fix, by delaying the pre-creation until the CVO reaches the
manifest-task node containing the ClusterOperator manifest.  That will
usually be the same node that has the other critical operator
manifests like the namespace, RBAC, and operator deployment.

Dropping fdef37dc0e's baremetal hack will re-expose us to issues on
install, where we race through all the manifests as fast as possible.
It's possible that we will now pre-create the ClusterOperator early
(because it's only blocked by the CRD) and still be a ways in front of
the operator pod coming up (because that needs a schedulable
control-plane node).  But we can address that by surpressing
ClusterOperatorDown and ClusterOperatorDegraded for some portion of
install in follow-up work.

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=1929917
[2]: https://bugzilla.redhat.com/show_bug.cgi?id=1957775