install/0000_90_cluster-version-operator_02_servicemonitor: Soften ClusterOperatorDegraded
During install, the CVO has pushed manifests into the cluster as fast
as possible without blocking on "has the in-cluster resource leveled?"
since way back in b0b4902fce (clusteroperator: Don't block on failing
during initialization, 2019-03-11, #136).  That can lead to
ClusterOperatorDown and ClusterOperatorDegraded firing during install,
as we see in [1], where:

* ClusterOperatorDegraded started pending at 5:00:15Z [2].
* Install completed at 5:09:58Z [3].
* ClusterOperatorDegraded started firing at 5:10:04Z [2].
* ClusterOperatorDegraded stopped firing at 5:10:23Z [2].
* The e2e suite complained about [1]:

    alert ClusterOperatorDegraded fired for 15 seconds with labels: {... name="authentication"...} (open bug: https://bugzilla.redhat.com/show_bug.cgi?id=1939580)

ClusterOperatorDown is similar, but I'll leave addressing it to a
separate commit.  For ClusterOperatorDegraded, the degraded condition
should not be particularly urgent [4], so we should be find bumping it
to 'warning' and using 'for: 30m' or something more relaxed than the
current 10m.

This commit brings back

* fb5257d4be
  (install/0000_90_cluster-version-operator_02_servicemonitor: Soften
  ClusterOperatorDegraded, 2021-05-06, #554) and
* 92ed7f110c
  (install/0000_90_cluster-version-operator_02_servicemonitor: Update
  ClusterOperatorDegraded message to 30m, 2021-05-08, #556).

There are some conflicts, because I am not bringing back 90539f9a17
(pkg/cvo/metrics: Ignore Degraded for cluster_operator_up, 2021-04-26, #550).
But that one had its own conflicts in metrics.go [5], and the
conflicts with this commit were orthogonal context issues, so moving
this back to 4.7 first won't make it much harder to bring back #550
and such later on, if we decide to do that.

[1]: https://prow.ci.openshift.org/view/gs/origin-ci-test/logs/release-openshift-ocp-installer-e2e-aws-upi-4.8/1389436726862155776
[2]: https://promecieus.dptools.openshift.org/?search=https://prow.ci.openshift.org/view/gs/origin-ci-test/logs/release-openshift-ocp-installer-e2e-aws-upi-4.8/1389436726862155776
     group by (alertstate) (ALERTS{alertname="ClusterOperatorDegraded"})
[3]: https://gcsweb-ci.apps.ci.l2s4.p1.openshiftapps.com/gcs/origin-ci-test/logs/release-openshift-ocp-installer-e2e-aws-upi-4.8/1389436726862155776/artifacts/e2e-aws-upi/clusterversion.json
[4]: https://github.com/openshift/api/pull/916
[5]: https://github.com/openshift/cluster-version-operator/pull/550#issuecomment-856815344