pkg/cvo/internal/operatorstatus: Replace wait-for with single-shot "is it alive now?"
Like cc9292a061 (lib/resourcebuilder: Replace wait-for with
single-shot "is it alive now?", 2020-07-07, #400), but for
ClusterOperator.  In that commit message, I'd explain that we'd
continue on past imperfect (e.g. Progressing=True) but still
happy-enough Deployments and such and later block on the
ClusterOperator.  But there's really no need to poll the
ClusterOperator while we're blocked on it; we can instead fail that
sync loop, take the short cool-off break, and come in again with a new
pass at the sync cycle.

This will help reduce delays like [1] where a good chunk of the ~5.5
minute delay was waiting for the network ClusterOperator to become
happy.  If instead we give up on that sync cycle and start in again
with a fresh sync cycle, we would have been more likely to recreate
the ServiceMonitors CRD more quickly.

The downside is that we will now complain about unavailable, degraded,
and unleveled operators more quickly, without giving them time to
become happy before complaining in ClusterVersion's status.  This is
mitigated by most of the returned errors being UpdateEffectNone, which
we will render as "waiting on..." Progressing messages.  The
exceptions are mostly unavailable, which is a serious enough condition
that I'm fine complaining about it aggressively, and degraded, which
has the fail-after-interval guard keeping us from complaining about it
too aggressively.

I'm also shifting the "does not declare expected versions" check
before the Get call.  It's goal is still to ensure that we fail in CI
before shipping an operator with such a manifest, and there's no point
in firing off an API GET before failing on a guard that only needs
local information.

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=1927168#c2