pkg/cincinnati: Log the proxy/TLS settings used for upstream fetches
AppendCertsFromPEM is pretty generous about parsing failures [1]:

  It appends any certificates found to s and reports whether any
  certificates were successfully parsed.

which can mask errors when we succesfully parse at least one cert, but
fail to parse some subsequent cert.  Make that a bit easier to detect
by logging the number of certs we are using by the time we make the
upstream update service request.  It would be more useful to log "does
the root CA pool have a CI for the target (proxy) URI?", but that
requires more CertPool expertise than I have.  The subject count
should be sufficient to determine whether the expected PEM contents
are being completely loaded or not, if we have access to the relevant
ConfigMap data as well.

We also use the same proxy/TLS approach to configure signature
fetches, but the risk of breaking down in the pipeline leading to
those signature requests seems small, and the logging I'm adding here
will cover the cluster-config -> CVO transmission, so I haven't added
logging to that side.  This will leave us uncovered for folks who do
not configure a channel, since they may be relying on the signature
fetching, but won't be using the upstream polling.  If we end up
worried about that, we can add logging for them in follow-up work.

I'm also logging the proxy URI host, if configured, because it's a
similar config pathway that folks might worry is not working.  I'm not
logging the full proxy URI, because the user information may include a
sensitive password.  Including only the host portion should be
sufficient to show that the configuration data path is working.

[1]: https://golang.org/pkg/crypto/x509/#CertPool.AppendCertsFromPEM