pkg/cvo/sync_worker: Increment Attempt on failed reconciliation
Avoid situations like [1]:

  $ grep 'Running sync.*in state\|Result of work' cvo.log | tail -n4
  I0517 11:30:37.342895       1 sync_worker.go:549] Running sync 4.8.0-0.nightly-2021-05-15-141455 (force=false) on generation 1 in state Reconciling at attempt 0
  I0517 11:36:19.258793       1 task_graph.go:555] Result of work: [Could not update servicemonitor "openshift-machine-api/cluster-autoscaler-operator" (255 of 676): the server does not recognize this resource, check extension API servers Could not update servicemonitor "openshift-kube-apiserver-operator/kube-apiserver-operator" (627 of 676): the server does not recognize this resource, check extension API servers update context deadline exceeded at 28 of 676 update context deadline exceeded at 28 of 676]
  I0517 11:39:38.268693       1 sync_worker.go:549] Running sync 4.8.0-0.nightly-2021-05-15-141455 (force=false) on generation 1 in state Reconciling at attempt 0
  I0517 11:45:20.182573       1 task_graph.go:555] Result of work: [Could not update servicemonitor "openshift-machine-api/cluster-autoscaler-operator" (255 of 676): the server does not recognize this resource, check extension API servers Could not update servicemonitor "openshift-kube-apiserver-operator/kube-apiserver-operator" (627 of 676): the server does not recognize this resource, check extension API servers]

where we are failing the reconcile, but not incrementing Attempt.
Because we rely on Attempt moving to drive changes in the ShiftOrder
and PermuteOrder calls in SyncWorker.apply.  This fixes a bug we've
had since the permuation logic landed in 1f95ccfa05 (sync: During
reconcile, try different parts of the payload and don't hang,
2019-04-17, #166).

As a side benefit, this change will no longer call
InitCOUpdateStartTime in SyncWorker.syncOnce after a failed
reconciliation cycle, so if a ClusterOperator goes Degraded during
reconciliation, we will now complain about it after the
UpdateEffectFailAfterInterval threshold elapses.

I'm a bit suspicious about 'steps := 8' capping ShiftOrder, because we
have some manifest task-nodes that have a good deal more than that.
But blindly increasing it would reduce the amount of permuting
PermuteOrder was doing.  We might have to do something more drastic
for reconciliation mode, like collecting manifest failures and
eventually failing the sync cycle, but not aborting it early to report
that failure.  But one step at a time; I'm not going to overhaul all
of that in this commit.

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=1927168#c10