pkg/cvo/sync_worker: Consolidate all ClusterOperator errors by reason
newClusterOperatorsNotAvailable is from c2ac20fa17 (status: Report the
operators that have not yet deployed, 2019-04-09, #158).  And the
not-available filtering is from bdd4545b1d (status: Hide generic
operator status in favor of more specific errors, 2019-05-19, #192).
But in ce1eda1095 (pkg/cvo/internal/operatorstatus: Change nested
message, 2021-02-04, #514), we moved "waiting on status.versions" from
ClusterOperatorNotAvailable to ClusterOperatorUpdating.  And we want
to avoid uncollapsed errors like:

  Multiple errors are preventing progress:
  * Cluster operator machine-api is updating versions
  * Cluster operator openshift-apiserver is updating versions

where we are waiting on multiple ClusterOperator which are in similar
situations.  This commit drops the filtering, because cluster
operators are important.  It does sort those errors to the end of the
list though, so the first error is the non-ClusterOperator error.

TestCVO_ParallelError no longer tests the consolidated error message,
because the consolidation is now restricted to ClusterOperator
resources.  I tried moving the
pkg/cvo/testdata/paralleltest/release-manifests manifests to
ClusterOperator, but then the test struggled with:

  I0802 16:04:18.133935    2005 sync_worker.go:945] Unable to precreate resource clusteroperator

so now TestCVO_ParallelError is excercising the fact that
non-ClusterOperator failures are not aggregated.